+++ kubernetes-examples/src/main/java/io/fabric8/openshift/examples/LoginExample.java
@@ -15,8 +15,10 @@
  */
 package io.fabric8.openshift.examples;
 
+import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
+import io.fabric8.openshift.api.model.Project;
 import io.fabric8.openshift.api.model.ProjectList;
 import io.fabric8.openshift.client.OpenShiftClient;
 import org.slf4j.Logger;
@@ -26,16 +28,15 @@ public class LoginExample {
   private static final Logger logger = LoggerFactory.getLogger(LoginExample.class);
 
   public static void main(String[] args) {
-
     try (DefaultKubernetesClient kubernetesClient = new DefaultKubernetesClient(new ConfigBuilder()
       .withMasterUrl("cluster_url")
       .withUsername("my_username")
       .withPassword("my_password")
       .build())) {
-
       final OpenShiftClient openShiftClient = kubernetesClient.adapt(OpenShiftClient.class);
+      logger.info("Login Successful");
       final ProjectList pl = openShiftClient.projects().list();
-      logger.info("{}", pl);
+      pl.getItems().stream().map(Project::getMetadata).map(ObjectMeta::getName).forEach(logger::info);
     }
   }
 }