+++ extensions/knative/generator/cmd/generate/generate.go
@@ -140,7 +140,7 @@ func main() {
 	manualTypeMap := map[reflect.Type]string{
 		reflect.TypeOf(apis.URL{}):             "java.lang.String",
 		reflect.TypeOf(apis.VolatileTime{}):    "java.lang.String",
-		reflect.TypeOf(runtime.RawExtension{}): "io.fabric8.kubernetes.api.model.HasMetadata",
+		reflect.TypeOf(runtime.RawExtension{}): "java.util.Map<String, Object>",
 	}
 
 	json := schemagen.GenerateSchema("http://fabric8.io/knative/KnativeSchema#", crdLists, providedPackages, manualTypeMap, packageMapping, mappingSchema, providedTypes, constraints)