+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Fix #2748: Pass custom headers in kubernetes-client to watch api by modify WatchConnectionManager
 * Fix #2745: Filtering Operations can't configure PropagationPolicy
 * Fix #2672: WaitUntilReady for Service resource throws IllegalArgumentException
+* Fix: Knative model genrator is now aligned with all other generators
 
 #### Improvements
 * Fix #2717: Remove edit() methods from RawCustomResourceOperationsImpl taking InputStream arguments